+++ tests/main_tests/contiguous_tests.rs
@@ -22,7 +22,7 @@ unsafe impl Contiguous for Tiny {
 }
 
 #[cfg(feature = "debug_checks")]
-#[repr(i8)]
+#[repr(i32)]
 #[derive(Debug, PartialEq, Copy, Clone)]
 enum Wrong {
     N1 = -1,
@@ -52,6 +52,42 @@ fn contiguous_accessors() {
     }
 }
 
+#[cfg(feature = "debug_checks")]
+#[repr(transparent)]
+#[derive(Debug, PartialEq, Copy, Clone)]
+struct SwappedLimits(u8);
+
+#[cfg(feature = "debug_checks")]
+unsafe impl Contiguous for SwappedLimits {
+    type Int = i8;
+
+    const MIN_VALUE: i8 = 2;
+    const MAX_VALUE: i8 = 0;
+}
+
+#[cfg(feature = "debug_checks")]
+#[test]
+fn swapped_limits() {
+    macro_rules! make_ic {
+        ($ty:ty) => {
+            ImplsContiguous::<SwappedLimits, $ty>::new_unchecked(
+                SwappedLimits::MIN_VALUE as _,
+                SwappedLimits::MAX_VALUE as _,
+            )
+        };
+    }
+    let ic = ImplsContiguous::<SwappedLimits, _>::NEW;
+    assert_eq!(ic.min_value(), &2);
+    assert_eq!(ic.max_value(), &0);
+
+    unsafe {
+        must_panic(|| drop(contiguous::from_i8(0, make_ic!(i8)))).unwrap();
+        must_panic(|| drop(contiguous::from_u8(0, make_ic!(u8)))).unwrap();
+        must_panic(|| drop(contiguous::from_i16(0, make_ic!(i16)))).unwrap();
+        must_panic(|| drop(contiguous::from_u16(0, make_ic!(u16)))).unwrap();
+    }
+}
+
 #[test]
 fn custom_type_tests() {
     for outside in (i8::MIN..=-2).chain(3..=i8::MAX) {
@@ -65,9 +101,18 @@ fn custom_type_tests() {
     }
 
     #[cfg(feature = "debug_checks")]
-    {
-        must_panic(|| drop(contiguous::from_i16::<Wrong>(0, infer!()))).unwrap();
-        must_panic(|| drop(contiguous::into_integer(Wrong::Z, infer!()))).unwrap();
+    unsafe {
+        macro_rules! make_ic {
+            ($ty:ty) => {
+                ImplsContiguous::<Wrong, $ty>::new_unchecked((Wrong::N1 as $ty).min(0), 1)
+            };
+        }
+        // making sure to test u8 since `from_u8` is manually written,
+        // while the rest are macro generated
+        must_panic(|| drop(contiguous::from_u8::<Wrong>(0, make_ic!(u8)))).unwrap();
+        must_panic(|| drop(contiguous::from_i8::<Wrong>(0, make_ic!(i8)))).unwrap();
+        must_panic(|| drop(contiguous::from_i16::<Wrong>(0, make_ic!(i16)))).unwrap();
+        must_panic(|| drop(contiguous::into_integer(Wrong::Z, make_ic!(i16)))).unwrap();
     }
 }